Support corner cases for GET and DELETE, return location on POST
- Return the location of the created resource in the "Location" response
header of POST /pet requests.
- Handle 400, 404 cases for GET and DELETE
- Add unit tests

+++ getapet/src/main/java/controller/PetController.java
@@ -1,6 +1,10 @@
 package controller;
 
+import java.net.URI;
+import java.net.URISyntaxException;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import error.ErrorResponseBody;
 import repository.PetRepository;
@@ -35,22 +40,40 @@ public class PetController {
 			consumes= {MediaType.APPLICATION_JSON_UTF8_VALUE, 
 							MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<?> addPet(@RequestBody @Validated Pet pet) {
-		pet.setResourceId(counterService.getNextSequence("pet"));
+		long resourceId = counterService.getNextSequence("pet");
+		pet.setResourceId(resourceId);
 		repository.save(pet);
-	    return new ResponseEntity<>(HttpStatus.CREATED);
+		
+		// Return resource URL in location header
+		HttpHeaders httpHeaders = new HttpHeaders();
+		httpHeaders.setLocation(ServletUriComponentsBuilder
+				.fromCurrentRequest().path("/{petId}")
+				.buildAndExpand(resourceId).toUri());
+		
+	    return new ResponseEntity<>(null, httpHeaders, HttpStatus.CREATED);
 	}
 	
 	@RequestMapping(method = RequestMethod.GET, value= "/{petId}", 
 			produces= {MediaType.APPLICATION_JSON_UTF8_VALUE, 
 							MediaType.APPLICATION_JSON_VALUE})
-	public Pet getPetById(@PathVariable Long petId) {
-	    return repository.findOne(petId);
+	public ResponseEntity<Pet> getPetById(@PathVariable Long petId) {
+	    Pet pet = repository.findOne(petId);
+	    if (pet == null) {
+	    	return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+	    }
+	    
+	    return new ResponseEntity<Pet>(pet, HttpStatus.OK);
 	}
 	
 	@RequestMapping(method = RequestMethod.DELETE, value= "/{petId}")
 	public ResponseEntity<?> deletePetById(@PathVariable Long petId) {
+		Pet pet = repository.findOne(petId);
+		if (pet == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
 		repository.delete(petId);
-	    return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 }
 
@@ -65,10 +88,13 @@ class PetControllerAdvice {
 
     @ResponseBody
     @ExceptionHandler(
-    	org.springframework.web.bind.MethodArgumentNotValidException.class
+    	{
+    		org.springframework.web.bind.MethodArgumentNotValidException.class,
+    		org.springframework.web.method.annotation.MethodArgumentTypeMismatchException.class
+    	}
     )
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorResponseBody httpBadRequestHandler(MethodArgumentNotValidException ex) {
+    ErrorResponseBody httpBadRequestHandler(Exception ex) {
         return new ErrorResponseBody(ex.getMessage());
     }
 }
